👷 퍼블리시 설정 공통화.
