+++ build.gradle.kts
@@ -1,4 +1,5 @@
 import dev.iurysouza.modulegraph.Orientation
+import groovy.json.JsonSlurper
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
@@ -9,6 +10,49 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.ksp) apply false
     alias(libs.plugins.modulegraph)
+}
+
+val configuration = JsonSlurper().parse(File(rootProject.projectDir, "configuration.json")) as Map<*, *>
+
+subprojects {
+    afterEvaluate {
+        if (plugins.hasPlugin(libs.plugins.android.library.get().pluginId) && true == configuration["PUBLISH"]) {
+            apply(plugin = "maven-publish")
+
+            extensions.getByType<PublishingExtension>().run {
+                repositories {
+                    maven {
+                        name = "LulKr"
+                        url = uri("https://maven.pkg.github.com/JustBurrow/packages")
+                        credentials {
+                            username = configuration["PUBLISH_PACKAGES_USER"] as String?
+                                ?: System.getenv("PUBLISH_PACKAGES_USER")
+                            password = configuration["PUBLISH_PACKAGES_TOKEN"] as String?
+                                ?: System.getenv("PUBLISH_PACKAGES_TOKEN")
+                        }
+                    }
+                }
+
+                publications {
+                    register<MavenPublication>("release") {
+                        groupId = "kr.lul.andoird.ui"
+                        artifactId = project.name.lowercase()
+                        version = libs.versions.ui.get()
+
+                        pom {
+                            scm {
+                                url = "https://github.com/JustBurrow/android-ui"
+                            }
+                        }
+
+                        afterEvaluate {
+                            from(components["release"])
+                        }
+                    }
+                }
+            }
+        }
+    }
 }
 
 moduleGraphConfig {
