+++ compose/build.gradle.kts
@@ -2,6 +2,8 @@ plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.compose.compiler)
     alias(libs.plugins.kotlin.android)
+
+    `maven-publish`
 }
 
 android {
@@ -29,6 +31,18 @@ android {
 
     kotlinOptions {
         jvmTarget = "17"
+    }
+
+    @Suppress("UnstableApiUsage")
+    testOptions {
+        unitTests.isReturnDefaultValues = true
+    }
+
+    publishing {
+        multipleVariants {
+            allVariants()
+            withJavadocJar()
+        }
     }
 }
